requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-cuser==2017.3.16
django-filter==25.1
django-jazzmin==3.0.1
django-rest-framework==0.1.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
environs==14.1.1
Flask==3.1.0
flask-cors==5.0.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
modeltranslation==0.25
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
Werkzeug==3.1.3
